Create makeScaleBarLabel util function
