+++ src/util/test/versionHandlers.test.ts
@@ -1,0 +1,26 @@
+import { expect } from "chai";
+
+import { makeScaleBarLabel } from "../versionHandlers";
+
+describe("Version handlers", () => {
+    describe("makeScaleBarLabel", () => {
+        // it("takes a string and converts it to sentence case", () => {
+        //     const startingString = "all lowercase. all lowercase";
+        //     expect(convertToSentenceCase(startingString)).to.equal(
+        //         "All lowercase. All lowercase"
+        //     );
+        // });
+        // it("ignores camelcase", () => {
+        //     const startingString = "all lowercase. has camelCase";
+        //     expect(convertToSentenceCase(startingString)).to.equal(
+        //         "All lowercase. Has camelCase"
+        //     );
+        // });
+        // it("fixes mid sentence caps", () => {
+        //     const startingString = "all lowercase. has Mid sentence cap";
+        //     expect(convertToSentenceCase(startingString)).to.equal(
+        //         "All lowercase. Has mid sentence cap"
+        //     );
+        // });
+    });
+});
